modulize dnxmy.py
