+++ dnxmy/__init__.py
@@ -1,0 +1,5 @@
+from .dnmmy import *
+from .config_genarator import *
+from .variable_generator import *
+
+__version__ = '0.0.0'
